+++ app/code/local/Stuntcoders/Banner/Model/Resource/Banner/Collection.php
@@ -1,7 +1,7 @@
 <?php
 
 
-class Stuntcoders_Banner_Model_Mysql4_Banner_Collection extends Mage_Core_Model_Mysql4_Collection_Abstract
+class Stuntcoders_Banner_Model_Resource_Banner_Collection extends Mage_Core_Model_Resource_Db_Collection_Abstract
 {
     protected function _construct()
     {