+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased:
+### CHANGED:
+- Move module under proper namespace
+
 ## 1.1.0 — 2015-02-26
 ### ADDED:
 - Add sort order to banner