+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This projec
 ## Unreleased:
 ### CHANGED:
 - Move module under proper namespace
+- Use default admin route
 
 ## 1.1.0 — 2015-02-26
 ### ADDED: