+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file. This projec
 - Update banner template
 - Prevent errors when banner can not be loaded
 - Use image path to save an image
+- Refactor deprecated resource models
 
 ## 1.1.0 — 2015-02-26
 ### ADDED: