+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased:
+## 1.2.0 — 2015-05-06
 ### ADDED:
 - Adminhtml layout updates
 - Banner group block